+++ tests/php/Game/Action/BuildCastleTest.php
@@ -2,7 +2,7 @@
 
 namespace FreeElephants\HexoNardsTests\Game\Action;
 
-use FreeElephants\HexoNards\Game\Action\BaseCastle;
+use FreeElephants\HexoNards\Game\Action\BuildCastle;
 use FreeElephants\HexoNards\Game\Army;
 use FreeElephants\HexoNards\Game\PlayerInterface;
 use FreeElephants\HexoNardsTests\AbstractTestCase;
@@ -10,7 +10,7 @@ use FreeElephants\HexoNardsTests\AbstractTestCase;
 /**
  * @author samizdam <samizdam@inbox.ru>
  */
-class BaseCastleTest extends AbstractTestCase
+class BuildCastleTest extends AbstractTestCase
 {
 
     public function testExecuteSuccess()
@@ -21,7 +21,7 @@ class BaseCastleTest extends AbstractTestCase
         $garrison->method('getOwner')->willReturn($player);
         $tile->setArmy($garrison);
 
-        $command = new BaseCastle($tile);
+        $command = new BuildCastle($tile);
         $command->execute($player);
 
         $this->assertTrue($tile->hasCastle());