+++ tests/php/Game/Action/ReplenishGarrisonTest.php
@@ -3,7 +3,7 @@
 namespace FreeElephants\HexoNardsTests\Game\Action;
 
 use FreeElephants\HexoNards\Game\Action\Exception\TouchForeignOwnException;
-use FreeElephants\HexoNards\Game\Action\ReplenishArmy;
+use FreeElephants\HexoNards\Game\Action\ReplenishGarrison;
 use FreeElephants\HexoNards\Game\Army;
 use FreeElephants\HexoNards\Game\PlayerInterface;
 use FreeElephants\HexoNardsTests\AbstractTestCase;
@@ -11,7 +11,7 @@ use FreeElephants\HexoNardsTests\AbstractTestCase;
 /**
  * @author samizdam <samizdam@inbox.ru>
  */
-class ReplenishArmyTest extends AbstractTestCase
+class ReplenishGarrisonTest extends AbstractTestCase
 {
 
     public function testExecuteSuccess()
@@ -19,7 +19,7 @@ class ReplenishArmyTest extends AbstractTestCase
         $player = $this->createMock(PlayerInterface::class);
         $army = new Army($player, $this->createTileWithMocks(), 10);
 
-        $command = new ReplenishArmy($army, 2);
+        $command = new ReplenishGarrison($army, 2);
         $command->execute($player);
 
         $this->assertCount(12, $army);
@@ -31,7 +31,7 @@ class ReplenishArmyTest extends AbstractTestCase
         $otherPlayer = $this->createMock(PlayerInterface::class);
         $army = new Army($otherPlayer, $this->createTileWithMocks(), 10);
 
-        $command = new ReplenishArmy($army, 2);
+        $command = new ReplenishGarrison($army, 2);
 
         $this->expectException(TouchForeignOwnException::class);
         $command->execute($player);